% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_JCHSlabels.R
\name{add_JCHSlabels}
\alias{add_JCHSlabels}
\title{Add JCHS variable labels to ACS files}
\usage{
add_JCHSlabels(df)
}
\arguments{
\item{df}{A dataframe created from load_acs of a single ACS year.}
}
\value{
Dataframe with ACS factor variables labeled.
}
\description{
Add JCHS variable labels to household plus file.
}
\note{
Labels are for standard JCHS variables only. Use add_ACSlabels for
other ACS labels. Created for 2023; check variables for changes before
using for prior years.
}
\examples{
load_acs(2023)
acs2023_l <- add_JCHSlabels(acs2023)
acs2023_l |> filter(as.integer(tenurecat) < 3) |> group_by(tenurecat) |> summarise(tot = sum(wgtp))

}
