% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_permits.R
\name{county_permits}
\alias{county_permits}
\title{Download county permits}
\usage{
county_permits(yr)
}
\arguments{
\item{yr}{The year to extract.}
}
\value{
A tibble with 31 columns.
}
\description{
Cleans census URLs to directly download and format annual county permit
datasets
}
\note{
Adapted from Chris Goodman:
"https://gist.github.com/cbgoodman/8b78153e93148cdff370e3c6f3629ade.js"
Can extract multiple years at once using purrr::map (see examples).
Recommend running rm(multiyear) once years of interest are extracted.
}
\examples{
cp23 <- county_permits(2023)

}
