% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_table.R
\name{export_table}
\alias{export_table}
\title{Export tables to workbook}
\usage{
export_table(input_wb, output_wb, tab)
}
\arguments{
\item{input_wb}{The .xlsx workbook to read, including the file path, in quotations.}

\item{output_wb}{The .xlsx workbook to write to, including the file path, in quotations.
If the same as the input_wb, the input_wb will be overwritten or added to.}

\item{tab}{The named tibble or object to write.}
}
\value{
Workbook with sheets named after input tab.
}
\description{
Exports tibbles and named objects to a worksheet of the same name in either
a new or existing workbook
}
\note{
Existing tab will be overwritten if worksheet already exists.
}
\examples{
tenTable <- acs2023 |> tab(tensimp, w = wgtp)
export_table("C:/Data/tabswb.xlsx", "C:/Data/tabswb.xlsx", tenTable)

}
