% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factcheck_format.R
\name{factcheck_format}
\alias{factcheck_format}
\title{Start factchecking spreadsheet}
\usage{
factcheck_format(txtfile)
}
\arguments{
\item{txtfile}{The full path and name of the text file in quotations.}
}
\value{
A dataframe where each line is a sentence or header.
}
\description{
Splits chapter at sentence breaks and puts each new sentence on a new row
}
\note{
Word files should be saved to .txt using Unicode (UTF-8) encoding.
Splits based on occurrence of ". " or a carriage return. Can use with
jointcenteR::export_table() to export to its own tab of a spreadsheet.
}
\examples{
chapter1 <- factcheck_format("C:/Data/testtext.txt")

}
