% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_permits.R
\name{get_permits}
\alias{get_permits}
\title{Get Building Permits Survey}
\usage{
get_permits(geography, year)
}
\arguments{
\item{geography}{Select the geography to pull. Valid inputs are "state" or
"county"}

\item{year}{The year to extract, starting with 1980 for states and 1990 for
counties.}
}
\value{
A tibble with 31 columns.
}
\description{
Cleans census URLs to directly download and format annual county or state permit
datasets
}
\note{
Adapted from Chris Goodman:
"https://gist.github.com/cbgoodman/8b78153e93148cdff370e3c6f3629ade.js"
Can extract multiple years at once using purrr::map (see examples).
}
\examples{
# pull state permits for 2023 and select unit variables
sp23 <- get_permits("state", 2023)
sp23 |> 
  select(year, region, state_name, starts_with("units_") & ends_with("unit"))

# pull multiple years at once and bind into long df 
sp_list <- map(2003:2023, get_permits, geography = "state")
sp <- bind_rows(sp_list)

}
