% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_acs.R
\name{load_acs}
\alias{load_acs}
\title{Load annual ACS household plus files}
\usage{
load_acs(yr, currentyear = 2023, path = acspath)
}
\arguments{
\item{yr}{The year of the data file to load in}

\item{currentyear}{The most recent year of data available, used for coding variables that don't need to be inflated.}

\item{path}{A path to the location of ACS .csv files.}
}
\value{
Dataframes loaded into global environment named with convention acs2023.
}
\description{
Reads annual ACS household plus files into global environment
}
\note{
Annual ACS files should be in one location and saved as .csv files.
}
\examples{
acspath <- "C:/Data/ACS/"
load_acs(2022, currentyear = 2023, path = acspath)

}
