% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_met.R
\name{load_met}
\alias{load_met}
\title{Load annual ACS metro files}
\usage{
load_met(yr, path = acspath)
}
\arguments{
\item{yr}{The year of the data file to load in}

\item{path}{A path to the location of metro .csv files.}
}
\value{
Dataframes loaded into global environment named with convention
met2023.
}
\description{
Reads annual ACS metro files into global environment
}
\note{
Annual metro files should be in one location and saved as .csv files. Load
multiple years with purrr::map.
}
\examples{
acspath <- "C:/Data/ACS/"
# single year
load_met(2022, path = acspath)

# multiple years
purrr::map(c(2019, 2023), load_met, path = acspath)

}
