% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mult.R
\name{load_mult}
\alias{load_mult}
\title{Load JCHS multiyear ACS file}
\usage{
load_mult(path = acspath, filename = multfilename)
}
\arguments{
\item{path}{A path to the location of multiyear .csv files.}

\item{filename}{The name of the file to load.}
}
\value{
Dataframes loaded into global environment named multiyear.
}
\description{
Reads JCHS multiyear ACS file into global environment
}
\note{
Multiyear files should be in one location and saved as .csv files.
Splitting the multiyear file for renters or owners will help them
load faster if you are routinely using one or the other.
}
\examples{
acspath <- "C:/Data/ACS/"
multfilename<- "ACS_multiyear_hhplus_RentersPlusVacant.csv"
load_mult()
load_mult(acspath, multfilename)

}
