% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_mult.R
\name{split_mult}
\alias{split_mult}
\title{Load JCHS multiyear ACS file}
\usage{
split_mult(yr, multdf = multiyear)
}
\arguments{
\item{yr}{The year to extract.}

\item{multdf}{The name of the dataframe to split. If using load_mult first,
the name should be multiyear.}
}
\value{
Dataframes loaded into global environment named with convention acs2001.
}
\description{
Reads JCHS multiyear ACS file into global environment
}
\note{
Can extract multiple years at once using purrr::map (see examples).
Recommend running rm(multiyear) once years of interest are extracted.
}
\examples{
acspath <- "C:/Data/ACS/"
multfilename<- "ACS_multiyear_hhplus_RentersPlusVacant.csv"

# split out multiple years
load_mult()
purrr::map(c(2001:2005), split_mult)

split_mult(2006)

}
