% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{One-way tables}
\usage{
tab(df, var, weight = unwgtd)
}
\arguments{
\item{df}{A dataframe containing the variable of interest and weight vector}

\item{var}{A categorical variable to tabulate.}

\item{weight}{A numeric weight variable. If no weight is provided, the tabulation is unweighted.}
}
\value{
A tibble of weighted or unweighted sums and shares.
}
\description{
Calculate weighted and weighted tables using one variable
}
\examples{
data("acs22")
# calculate the number and share of households by tenure using household weights
tab(acs22, ten, wgtp) 

#unweighted observations
tab(acs22, ten)

}
