% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab2.R
\name{tab2}
\alias{tab2}
\title{Two-way tables}
\usage{
tab2(df, var, var2, weight = unwgtd)
}
\arguments{
\item{df}{A dataframe containing the variable of interest and weight vector}

\item{var, var2}{A categorical variable to tabulate.}

\item{weight}{A numeric weight variable. If no weight is provided, the tabulation is unweighted.}
}
\value{
A tibble of weighted or unweighted sums and shares.
}
\description{
Calculate weighted and weighted tables using one variable
}
\note{
The first variable entered is the grouping variable for calculating shares.
}
\examples{
data("acs22")
# calculate the number and share of households by tenure and disability using household weights
tab2(acs22, ten, dis, wgtp) 

#unweighted observations
tab2(acs22, ten, dis)

}
