% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgtd_med.R
\name{wgtd_med}
\alias{wgtd_med}
\title{Single-variable medians}
\usage{
wgtd_med(df, var, weight = "unwgtd")
}
\arguments{
\item{df}{A dataframe containing the variable of interest and weight vector}

\item{var}{A continuous variable to tabulate.}

\item{weight}{A numeric weight variable. If no weight is provided, the tabulation is unweighted.}
}
\value{
A tibble of weighted or unweighted medians.
}
\description{
Calculate weighted and weighted medians using one variable
}
\examples{
data("acs22")
# calculate the weighted median household income
wgtd_med(acs22, hincp, wgtp) 

#returns unweighted median if no weight if provided
wgtd_med(acs22, hincp)

}
