% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgtd_med2.R
\name{wgtd_med2}
\alias{wgtd_med2}
\title{Two-variable medians}
\usage{
wgtd_med2(df, var, var2, weight = "unwgtd")
}
\arguments{
\item{df}{A dataframe containing the variable of interest and weight vector}

\item{var}{A categorical variable to group by.}

\item{var2}{A continuous variable to tabulate median.}

\item{weight}{A numeric weight variable. If no weight is provided, the tabulation is unweighted.}
}
\value{
A tibble of weighted or unweighted sums and shares.
}
\description{
Calculate weighted and weighted tables using one variable
}
\note{
The first variable entered is the grouping variable.
}
\examples{
data("acs22")
# calculate the weighted median household income by tenure
wgtd_med2(acs22, ten, hincp, wgtp) 

#returns unweighted medians if no weight if provided
wgtd_med2(acs22, ten, hincp)

}
