% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd_med.R
\name{wtd_med}
\alias{wtd_med}
\title{Grouped medians}
\usage{
wtd_med(df, var, var2 = NULL, weight = NULL)
}
\arguments{
\item{df}{A dataframe containing the variable(s) of interest and weight vector}

\item{var}{A continuous variable to tabulate median.}

\item{var2}{A categorical variable to group by.}

\item{weight}{A numeric weight variable. If no weight is provided,
the tabulation is unweighted.}
}
\value{
A tibble of weighted or unweighted medians.
}
\description{
Calculate weighted and weighted medians with up to one grouping variable
}
\note{
The second variable entered is the grouping variable. Does not
currently work in combination with the mutate or summarise function.
To calculate medians when this doesn't work, use
Hmisc::wtd.quantile(var, probs=0.5).
}
\examples{
data("acs22")
# calculate the weighted median household income by tenure
wgtd_med2(acs22, ten, hincp, wgtp) 

#returns unweighted medians if no weight if provided
wgtd_med2(acs22, ten, hincp)

}
