% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hvs.R
\name{get_hvs}
\alias{get_hvs}
\title{Get Housing Vacancy Survey data}
\usage{
get_hvs()
}
\value{
A tibble of HVS variables for each quarter from 2Q2001 through most
recent available.
}
\description{
Pulls and organizes Housing Vacancy Survey data using tidyquant
}
\note{
Adapted from Len Kiefer:
http://lenkiefer.com/2017/09/18/a-tidyquant-um-of-solace/
}
\examples{
hvs <- get_hvs()

}
