% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nrc.R
\name{get_nrc}
\alias{get_nrc}
\title{Get New Residential Construction data}
\usage{
get_nrc(type = "monthly")
}
\arguments{
\item{type}{Specifies the interval of interest, either "monthly" or "annual"}
}
\value{
A tibble of HVS variables for each quarter from 2Q2001 through most
recent available.
}
\description{
Pulls and organizes New Residential Construction data using tidyquant
}
\note{
Adapted from Len Kiefer:
http://lenkiefer.com/2017/09/18/a-tidyquant-um-of-solace/
Annual counts are totals of monthly NSA values, except annual under
construction counts are the value in December of a given year. If
current year is incomplete, it is omitted from annual table. Monthly
counts are seasonally adjusted annual rates, except under construction
counts are the seasonally adjusted rate (not annualized).
}
\examples{
# pull annual data
nrc_a <- get_nrc(type = "annual") |> 
  # keep vars for year and units in buildings with 2+ units
  select(year, contains('_2pl')) |> 
  # keep years in range of interest
  filter(year >= 2000)

# pull monthly data to calculate ytd SA average
nrc_m <- get_nrc("monthly") |> 
  # reshape to make calculations easier
  pivot_longer(!c("year", "month"), names_to = "var", values_to = "value") |>
  filter(str_detect(var, '_2pl') & year == 2024) |> # keep current year & mf vars
  group_by(year, var) |>   
  mutate(avg = mean(value)) |>   
  filter(month == 1) |>   
  select(-month, -value) |>   
  pivot_wider(names_from = var, values_from = avg)

#combine annual and ytd dfs  
nrc <- rbind(nrc_a, nrc_m)

}
