% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflate_ai.R
\name{inflate_ai}
\alias{inflate_ai}
\title{Inflate with CPI-U All Items}
\usage{
inflate_ai(
  initial_amount,
  reference_year,
  target_year,
  fred_api_key = Sys.getenv("FRED_API_KEY")
)
}
\arguments{
\item{initial_amount}{A numeric value or vector to be inflated.}

\item{reference_year}{The year the values to be inflated are from.}

\item{target_year}{The year values should be inflated to.}

\item{fred_api_key}{Your unique FRED API key.}
}
\value{
A numeric vector of inflated values.
}
\description{
Inflate values from reference to target dollars using CPI-U All Items, annual
NSA
}
\note{
CPI-U inflation rates are pulled from FRED using fredr.
Monthly values are averaged for the year. Set your unique FRED API key here:
https://fredaccount.stlouisfed.org/apikeys and add it to your R global
environment using Sys.setenv(FRED_API_KEY = '{put your key here}').
}
\examples{
# inflate value from 2000 dollars to 2023 dollars
inflate_ai(5000, 2000, 2023)

data("acs22")
# inflate household incomes to target year dollars in new variable
acs22 |> mutate(hincp_infl_23 = inflate_ai(hincp, 2022, 2023))

}
