% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointcenteR-package.R
\docType{package}
\name{jointcenteR-package}
\alias{jointcenteR}
\alias{jointcenteR-package}
\title{jointcenteR: Perform routine Joint Center tasks}
\description{
Loads common datasets into global environmnet. Splits multiyear ACS file into manageable dataframes. Calculates weighted and unweighted tables with counts and shares or medians for large survey datasets. Inflates continuous variables to target year dollars using CPI-U All Items, and inflates rents to target year dollars using CPI-U Less Shelter. Exports tabs to an excel workbook.
}
\author{
\strong{Maintainer}: Whitney Airgood-Obrycki \email{whitney_airgood-obrycki@harvard.edu} (\href{https://orcid.org/0000-0003-3644-1716}{ORCID})

}
\keyword{internal}
