% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_acs.R
\name{load_acs}
\alias{load_acs}
\title{Load annual ACS household plus files}
\usage{
load_acs(year, path = acspath)
}
\arguments{
\item{year}{The year of the data file to load in}

\item{path}{A path to the location of ACS .csv files.}
}
\value{
Dataframes loaded into global environment named with convention
acs2023.
}
\description{
Reads annual ACS household plus files into global environment
}
\note{
Annual ACS files should be in one location and saved as .csv files. Load
multiple years with purrr::map.
}
\examples{
acspath <- "C:/Data/ACS/"
# single year
load_acs(2022, path = acspath)

# multiple years
purrr::map(c(2019, 2021:2023), load_acs, path = acspath)

}
