% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{Count and share tables}
\usage{
tab(df, var, var2 = NULL, var3 = NULL, weight = NULL)
}
\arguments{
\item{df}{A dataframe containing the variable of interest and weight vector}

\item{var, var2, var3}{A categorical variable to tabulate.}

\item{weight}{A numeric weight variable. If no weight is provided, the
tabulation is unweighted.}
}
\value{
A tibble of weighted or unweighted sums and shares.
}
\description{
Calculate weighted and weighted tables using up to three variables
}
\note{
The first two variables entered are the grouping variables for
calculating shares.
}
\examples{
data("acs22")
# calculate the number and share of households by region, tenure, and
disability using household weights
tab3(acs22, region, ten, dis, wgtp)

# unweighted observations
tab3(acs22, region, ten, dis)

}
